Increase file descriptor limit in native AOT executables
Fixes #82719
