+++ src/coreclr/nativeaot/Bootstrap/main.cpp
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 #include <stdint.h>
+#include <minipal/descriptorlimit.h>
 
 #if defined(DEBUG) && defined(_WIN32)
 #include <process.h>
@@ -222,6 +223,8 @@ int main(int argc, char* argv[])
     int initval = InitializeRuntime();
     if (initval != 0)
         return initval;
+
+    minipal_increase_descriptor_limit();
 
 #if defined(DEBUG) && defined(_WIN32)
     // work around Debug UCRT shutdown issues: https://github.com/dotnet/runtime/issues/108640
