+++ src/mono/CMakeLists.txt
@@ -857,8 +857,6 @@ set(FULL_VERSION ${product_version_string})
 ######################################
 
 if(CLR_CMAKE_TARGET_LINUX_MUSL)
-  # Setting RLIMIT_NOFILE breaks debugging of coreclr on musl-libc for some reason
-  add_definitions(-DDONT_SET_RLIMIT_NOFILE)
   # On musl-libc, we need to ensure that the reported stack range for the primary thread is
   # larger than the initial committed stack size.
   add_definitions(-DENSURE_PRIMARY_STACK_SIZE)
